feat: extract feature module
