+++ packages/backend/server/src/modules/quota/types.ts
@@ -1,40 +1,20 @@
-import type { Prisma } from '@prisma/client';
-
-export enum FeatureKind {
-  Feature,
-  Quota,
+export enum QuotaType {
+  Quota_FreePlanV1 = 'free_plan_v1',
+  Quota_ProPlanV1 = 'pro_plan_v1',
 }
-
-export type CommonFeature = {
-  feature: FeatureType;
-  type: FeatureKind;
-  version: number;
-  configs: Prisma.InputJsonValue;
-};
-
-export type Feature = CommonFeature & {
-  type: FeatureKind.Feature;
-};
 
 export type Quota = CommonFeature & {
   type: FeatureKind.Quota;
+  feature: QuotaType;
   configs: {
     blobLimit: number;
     storageQuota: number;
   };
 };
 
-export enum FeatureType {
-  // features
-  Feature_EarlyAccess = 'early_access',
-  // quotas
-  Quota_FreePlanV1 = 'free_plan_v1',
-  Quota_ProPlanV1 = 'pro_plan_v1',
-}
-
 export const Quotas: Quota[] = [
   {
-    feature: FeatureType.Quota_FreePlanV1,
+    feature: QuotaType.Quota_FreePlanV1,
     type: FeatureKind.Quota,
     version: 1,
     configs: {
@@ -45,7 +25,7 @@ export const Quotas: Quota[] = [
     },
   },
   {
-    feature: FeatureType.Quota_ProPlanV1,
+    feature: QuotaType.Quota_ProPlanV1,
     type: FeatureKind.Quota,
     version: 1,
     configs: {
